constexpr Camera
