+++ src/core/CMakeLists.txt
@@ -1,5 +1,4 @@
 target_sources(ufpslib PUBLIC
-  camera.cpp
   entity.cpp
   render_entity.cpp
 )
